requirements.txt
django==1.4
python-ldap==2.4.3
django-auth-ldap==1.1
PIL==1.1.7
South==0.7.3
sorl-thumbnail
django-bootstrap-toolkit


# To use opera
 MySQL-python 